pyproject.toml
[tool.poetry]
name = "ghostsnstuff-spiritbox-fw"
version = "0.1.0"
description = "Firmware for a Raspberry PI based spirit box connected to OpenAI."
authors = ["LifetimeMistake"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.52.1"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
